clean up batch scrobbling

+++ iOS/terrible-player/api/managers/TBPLastFMTrackManager.m
@@ -170,7 +170,7 @@
 
 - (NSString *)scrobbleParamWithName:(NSString *)name index:(NSUInteger)index
 {
-    return [NSString stringWithFormat:@"%@[%u]", name, index];
+    return [NSString stringWithFormat:@"%@[%lu]", name, (unsigned long)index];
 }
 
 @end
